Make AggressiveAI respond to being hit
Since we don't know what hit it, we can increase the agro timer.  That
way if the enemy was not in an agro state, nearby enemies won't be set
to agro by one enemy being sniped, and you can avoid making it agro by
not going near it when it's in its aware state.